Move HTMLDocument members to Document.
